[refactor] 게시물 글쓰기 - 이미지 업로드 컴포넌트 분리
